+++ pkg/api/api.go
@@ -297,7 +297,7 @@ func (a *API) RegisterDistributor(d *distributor.Distributor, pushConfig distrib
 	a.RegisterRoute("/distributor/ring", d, false, true, "GET", "POST")
 	a.RegisterRoute("/distributor/all_user_stats", http.HandlerFunc(d.AllUserStatsHandler), false, true, "GET")
 	a.RegisterRoute("/distributor/ha_tracker", d.HATracker, false, true, "GET")
-	a.RegisterRoute("/distributor/start_artificial_latency/{latencyMin}/{latencyMax}/{latencyDuration}/{latencyWorkers}/{latencyErrorFrequency}", http.HandlerFunc(d.StartArtificialLatencyHandler), false, true, "POST")
+	a.RegisterRoute("/distributor/start_artificial_latency/{latencyMin}/{latencyMax}/{latencyDuration}/{latencyWorkers}/{latencySlowDownFrequency}/{latencyErrorFrequency}", http.HandlerFunc(d.StartArtificialLatencyHandler), false, true, "POST")
 	a.RegisterRoute("/distributor/stop_artificial_latency", http.HandlerFunc(d.StopArtificialLatencyHandler), false, true, "POST")
 }