+++ pkg/mimirpb/custom.go
@@ -452,6 +452,80 @@ type MarshalerWithSize interface {
 	MarshalWithSize(size int) ([]byte, error)
 }
 
+func metadataSetFromSettings(skipDeduplicateMetadata bool) metadataSet {
+	if skipDeduplicateMetadata {
+		return newPassthroughMetadataSet()
+	}
+	return newDedupingMetadataSet()
+}
+
+// metadataSet is the collection of metadata within a request.
+// It keeps the order at which metadata is added. Metadata may optionally be deduplicated by family name.
+type metadataSet interface {
+	add(family string, mm MetricMetadata)
+	len() int
+	slice() []*MetricMetadata
+}
+
+var _ metadataSet = dedupingMetadataSet{}
+var _ metadataSet = &passthroughMetadataSet{}
+
+// dedupingMetadataSet is a metadataSet that only stores one metadata per metric family.
+// Only the first metadata seen for a given family is kept.
+type dedupingMetadataSet struct {
+	deduplicated map[string]*orderAwareMetricMetadata
+}
+
+func newDedupingMetadataSet() dedupingMetadataSet {
+	return dedupingMetadataSet{
+		deduplicated: make(map[string]*orderAwareMetricMetadata),
+	}
+}
+
+func (m dedupingMetadataSet) add(family string, mm MetricMetadata) {
+	if _, ok := m.deduplicated[family]; ok {
+		// Already have metadata for this metric familiy name.
+		// Since we cannot have multiple definitions of the same
+		// metric family name, we ignore this metadata.
+		return
+	}
+	m.deduplicated[family] = &orderAwareMetricMetadata{MetricMetadata: mm, order: m.len()}
+}
+
+func (m dedupingMetadataSet) len() int {
+	return len(m.deduplicated)
+}
+
+func (m dedupingMetadataSet) slice() []*MetricMetadata {
+	result := make([]*MetricMetadata, m.len())
+	for _, meta := range m.deduplicated {
+		result[meta.order] = &meta.MetricMetadata
+	}
+	return result
+}
+
+type passthroughMetadataSet struct {
+	metadata []*MetricMetadata
+}
+
+func newPassthroughMetadataSet() *passthroughMetadataSet {
+	return &passthroughMetadataSet{
+		metadata: make([]*MetricMetadata, 0),
+	}
+}
+
+func (m *passthroughMetadataSet) add(family string, mm MetricMetadata) {
+	m.metadata = append(m.metadata, &mm)
+}
+
+func (m *passthroughMetadataSet) len() int {
+	return len(m.metadata)
+}
+
+func (m *passthroughMetadataSet) slice() []*MetricMetadata {
+	return m.metadata
+}
+
 // orderAwareMetricMetadata is a tuple (index, metadata) that knows its own position in a metadata slice.
 // It's tied to custom logic that unmarshals RW2 metadata into a map, and allows us to
 // remember the order that metadata arrived in when unmarshalling.