+++ pkg/mimirpb/prealloc_rw2.go
@@ -29,6 +29,7 @@ func FromWriteRequestToRW2Request(rw1 *WriteRequest, commonSymbols []string, off
 		SkipLabelCountValidation:        rw1.SkipLabelCountValidation,
 		skipUnmarshalingExemplars:       rw1.skipUnmarshalingExemplars,
 		skipNormalizeMetadataMetricName: rw1.skipNormalizeMetadataMetricName,
+		skipDeduplicateMetadata:         rw1.skipDeduplicateMetadata,
 	}
 
 	symbols := symbolsTableFromPool()